Deprecated DLModelBuilder: VGG16, VGG19
